+++ tensorrt_llm/_torch/utils.py
@@ -314,52 +314,3 @@ def get_device_uuid(device_idx: int) -> str:
     property = torch.cuda.get_device_properties(device_idx)
     uuid = "GPU-" + str(property.uuid)
     return uuid
-
-
-_buffer_pool = None
-
-
-def set_shared_pool(buffer_pool):
-    """Sets the global memory pool for buffer allocation.
-
-    Args:
-        buffer_pool: A CUDA memory pool object to use for allocations.
-    """
-    global _buffer_pool
-    _buffer_pool = buffer_pool
-
-
-def get_shared_pool():
-    """Retrieves the current global memory pool.
-
-    Returns:
-        The current memory pool, or None if not set.
-    """
-    return _buffer_pool
-
-
-@contextlib.contextmanager
-def with_shared_pool(buffer_pool) -> contextlib.AbstractContextManager:
-    """Temporarily sets a preferred memory pool and restores the previous one on exit.
-
-    This context manager allows temporarily switching to a different memory pool
-    for CUDA graph operations, ensuring the original pool is restored even if
-    an exception occurs.
-
-    Args:
-        mem_pool: The memory pool to use within the context.
-
-    Yields:
-        None
-
-    Example:
-        >>> with with_shared_pool(buffer_pool):
-        ...     # Allocations within this block use buffer_pool
-        ...     tensor = allocate_buffer(...)
-    """
-    old_buffer_pool = get_shared_pool()
-    set_shared_pool(mem_pool)
-    try:
-        yield
-    finally:
-        set_shared_pool(old_buffer_pool)
